Fix FCM entity ID type width + add ServerInfo.NexusId (#75)
* Fix entity ID type and add Nexus ID support in FCM notifications

* Map unknown FCM entity types to null instead of an undefined enum

ToEntityEvent cast body.EntityType (int?) straight to EntityType?, so an
out-of-range value (a future server type or malformed data) produced an
undefined enum member rather than null — defeating the type-safety the enum
is meant to provide and surprising consumers that switch over it. Validate
with Enum.IsDefined via a new ToEntityType helper; absent or unknown values
now map to null, matching how ParsePairingEntity already treats unknown types.


---------

+++ src/RustPlusApi.Fcm/Data/FcmMessage.cs
@@ -87,7 +87,7 @@ public sealed record Body
     public int? EntityType { get; init; }
 
     /// <summary>Entity ID when <see cref="Type"/> is <c>"entity"</c>.</summary>
-    public int? EntityId { get; init; }
+    public ulong? EntityId { get; init; }
 
     /// <summary>Entity name when <see cref="Type"/> is <c>"entity"</c>.</summary>
     public string? EntityName { get; init; }
